+++ .github/workflows/main.yml
@@ -1,6 +1,9 @@
 name: CI
 
-on: push
+on: 
+  push:
+   branches:
+      - ops-quard-rails
 
 jobs:
   build:
