--- src/index.test.js
@@ -1,5 +1,0 @@
-const sayHi = require('./index.js')
-
-test("Says 'Hello Mike' when 'Mike' is passed", () => {
-   expect(sayHi('Mike')).toBe('Hello Mike')
-})
